requirements.txt
psycopg2-binary
Django
django-rest-framework
django-filter